mackia_bot/requirements.txt
anyio==4.4.0
Brotli==1.1.0
certifi==2024.8.30
charset-normalizer==3.3.2
cryptg==0.4.0
ffmpeg-python==0.2.0
ffmpy==0.4.0
h11==0.14.0
hachoir==3.3.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
mutagen==1.47.0
pillow==10.4.0
pyaes==1.6.1
pyasn1==0.6.0
pycryptodomex==3.20.0
python-dotenv==1.0.1
python-telegram-bot==21.5
requests==2.32.3
rsa==4.9
setuptools==74.1.2
six==1.16.0
sniffio==1.3.1
Telethon==1.36.0
typing_extensions==4.12.2
urllib3==2.2.2
urwid==2.6.15
wcwidth==0.2.13
websockets==13.0.1
yt-dlp==2024.8.6
